+++ tests/HitObjectCompile/UnityEngineCompileStubs.cs
@@ -1,0 +1,28 @@
+namespace UnityEngine
+{
+    public class Object
+    {
+    }
+
+    public class GameObject : Object
+    {
+    }
+
+    public class Collision
+    {
+    }
+
+    public class MonoBehaviour : Object
+    {
+        protected static void Destroy(Object target)
+        {
+        }
+    }
+
+    public static class Debug
+    {
+        public static void Log(object message)
+        {
+        }
+    }
+}